# Accelerate is_inside function during cutcode optimisation