--- a/meerk40t/tools/pathtools.py
+++ b/meerk40t/tools/pathtools.py
@@ -729,6 +729,44 @@ class VectorMontonizer:
                 seg.active = True
                 self._actives.append(seg)
 
+    def add_pointlist(self, path):
+        """
+        Add segments in the form of a connected path. These positions are read and segments are created for these
+        points.
+
+        @param path:
+        @return:
+        """
+        self._dirty_scanline = True
+        self._dirty_event_sort = True
+        self._dirty_actives_sort = True
+        for i in range(len(path) - 1):
+            p0 = Point(path[i])
+            p1 = Point(path[i + 1])
+            if p0.y > p1.y:
+                high = p0
+                low = p1
+            else:
+                high = p1
+                low = p0
+
+            # b = low.y - (m * low.x)
+            if self.valid_low > high.y:
+                # Cluster before range.
+                continue
+            if self.valid_high < low.y:
+                # Cluster after range.
+                continue
+            seg = Segment(p0, p1)
+            # cluster = [False, i, p0, p1, high, low, m, b, path]
+            if self.valid_low < low.y:
+                self._events.append((low.y, seg))
+            if self.valid_high > high.y:
+                self._events.append((high.y, seg))
+            if high.y >= self.scanline >= low.y:
+                seg.active = True
+                self._actives.append(seg)
+
     def current_is_valid_range(self):
         return self.valid_high >= self.scanline >= self.valid_low