Tested a couple of algorithms and reused VectorMontizer with an adapted interpolation algorithm to better fit curves with acute slope changes

```
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.041s, ray-new=False in 0.000s, shapely=False in 0.000s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.017s, ray-new=False in 0.000s, shapely=False in 0.000s
Tolerance: 0.0, vm=False in 0.001s, sb=False in 0.001s, ray-old=False in 0.016s, ray-new=False in 0.001s, shapely=False in 0.001s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.001s, ray-new=False in 0.001s, shapely=False in 0.001s
Tolerance: 0.0, vm=True in 0.005s, sb=True in 0.005s, ray-old=True in 0.014s, ray-new=True in 0.016s, shapely=True in 0.010s
Tolerance: 0.0, vm=True in 0.004s, sb=True in 0.004s, ray-old=True in 0.001s, ray-new=True in 0.233s, shapely=True in 0.124s
Tolerance: 0.0, vm=False in 0.001s, sb=False in 0.001s, ray-old=False in 0.038s, ray-new=False in 0.003s, shapely=False in 0.003s
Tolerance: 0.0, vm=False in 0.002s, sb=False in 0.002s, ray-old=False in 0.012s, ray-new=False in 0.001s, shapely=False in 0.001s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.046s, ray-new=False in 0.000s, shapely=False in 0.000s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.020s, ray-new=False in 0.000s, shapely=False in 0.000s
Tolerance: 0.0, vm=False in 0.001s, sb=False in 0.001s, ray-old=False in 0.012s, ray-new=False in 0.001s, shapely=False in 0.001s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.001s, ray-new=False in 0.000s, shapely=False in 0.000s
Tolerance: 0.0, vm=True in 0.005s, sb=True in 0.005s, ray-old=True in 0.013s, ray-new=True in 0.013s, shapely=True in 0.013s
Tolerance: 0.0, vm=True in 0.002s, sb=True in 0.002s, ray-old=True in 0.001s, ray-new=True in 0.249s, shapely=True in 0.130s
Tolerance: 0.0, vm=False in 0.000s, sb=False in 0.000s, ray-old=False in 0.035s, ray-new=False in 0.002s, shapely=False in 0.003s
Tolerance: 0.0, vm=False in 0.002s, sb=False in 0.002s, ray-old=False in 0.014s, ray-new=False in 0.001s, shapely=False in 0.001s
```

## Summary by Sourcery

Enhancements:
- Improve the performance of the is_inside function by introducing a new raycasting algorithm and adapting the VectorMontonizer for better handling of curves with acute slope changes.